시나리오 조인 작성중, 하위 테이블 작성: 아바타, 월드파츠, 장르, 태그

+++ src/main/java/com/ohgiraffers/r_pakabe/domains/scenarioAvatars/command/domain/model/ScenarioAvatar.java
@@ -15,7 +15,7 @@ public class ScenarioAvatar {
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long senarioAvatarId;
+    private Integer senarioAvatarId;
 
     @Column
     @ColumnDefault("여기에 이름을 입력")
