+++ src/main/java/com/ohgiraffers/r_pakabe/domains/scenarioAvatars/command/domain/repository/ScenarioAvatarRepository.java
@@ -3,5 +3,5 @@ package com.ohgiraffers.r_pakabe.domains.scenarioAvatars.command.domain.reposito
 import com.ohgiraffers.r_pakabe.domains.scenarioAvatars.command.domain.model.ScenarioAvatar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScenarioAvatarRepository extends JpaRepository<ScenarioAvatar, Long> {
+public interface ScenarioAvatarRepository extends JpaRepository<ScenarioAvatar, Integer> {
 }
